made changes regarding routing in next and completed lesson four
