--- components/LinkList.tsx
@@ -1,35 +1,0 @@
-import React from "react";
-import Link from "./Link";
-import { ILink } from "../types/types";
-import { useQuery, gql } from "@apollo/client";
-
-const FEED_QUERY = gql`
-  {
-    feed {
-      id
-      links {
-        id
-        createdAt
-        url
-        description
-      }
-    }
-  }
-`;
-
-const LinkList = () => {
-  const { data } = useQuery(FEED_QUERY);
-  return (
-    <div>
-      {data && (
-        <>
-          {data.feed.links.map((link: ILink) => (
-            <Link key={link.id} link={link} />
-          ))}
-        </>
-      )}
-    </div>
-  );
-};
-
-export default LinkList;
