+++ pages/index.tsx
@@ -1,22 +1,55 @@
 import type { NextPage } from "next";
 import Head from "next/head";
 import Image from "next/image";
-import CreateLink from "../components/CreateLink";
-import LinkList from "../components/LinkList";
-import styles from "../styles/Home.module.css";
+import Header from "../components/Header";
+import Link from "../components/Link";
+import { ILink } from "../types/types";
+import { useQuery, gql } from "@apollo/client";
+
+const FEED_QUERY = gql`
+  {
+    feed {
+      id
+      links {
+        id
+        createdAt
+        url
+        description
+      }
+    }
+  }
+`;
+
+const LinkList = () => {
+  const { data } = useQuery(FEED_QUERY);
+  return (
+    <div>
+      {data && (
+        <>
+          {data.feed.links.map((link: ILink) => (
+            <Link key={link.id} link={link} />
+          ))}
+        </>
+      )}
+    </div>
+  );
+};
 
 const Home: NextPage = () => {
   return (
     <>
       <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
+        <title>Hacker News</title>
+        <meta
+          name="description"
+          content="news, hackernews, ondemand news, world news"
+        />
         <link rel="icon" href="/favicon.ico" />
       </Head>
 
-      <div className="main">
+      <div className="center w85">
+        <Header />
         <LinkList />
-        <CreateLink />
       </div>
     </>
   );
